PUBDEV-7739: add FI for GBM
