+++ h2o-core/src/main/java/hex/FeatureInteractions.java
@@ -1,5 +1,6 @@
 package hex;
 
+import hex.genmodel.algos.tree.SharedTreeNode;
 import org.apache.commons.lang.mutable.MutableInt;
 import water.util.TwoDimTable;
 
@@ -219,5 +220,88 @@ public class FeatureInteractions {
 
     public Set<Map.Entry<String, FeatureInteraction>> entrySet() {
         return map.entrySet();
+    }
+
+    public static void collectFeatureInteractions(SharedTreeNode node, List<SharedTreeNode> interactionPath,
+                                                  double currentGain, double currentCover, double pathProba, int depth, int deepening,
+                                                  FeatureInteractions featureInteractions, Set<String> memo, int maxInteractionDepth,
+                                                  int maxTreeDepth, int maxDeepening, int treeIndex) {
+
+        if (node.isLeaf() || depth == maxTreeDepth) {
+            return;
+        }
+
+        interactionPath.add(node);
+        currentGain += node.getSquaredError();
+        currentCover += node.getWeight();
+
+        double ppl = pathProba * (node.getLeftChild().getWeight() / node.getWeight());
+        double ppr = pathProba * (node.getRightChild().getWeight() / node.getWeight());
+
+        FeatureInteraction featureInteraction = new FeatureInteraction(interactionPath, currentGain, currentCover, pathProba, depth, 1, treeIndex);
+
+        if ((depth < maxDeepening) || (maxDeepening < 0)) {
+            collectFeatureInteractions(node.getLeftChild(), new ArrayList<>(), 0, 0, ppl, depth + 1,
+                    deepening + 1, featureInteractions, memo, maxInteractionDepth, maxTreeDepth, maxDeepening, treeIndex);
+            collectFeatureInteractions(node.getRightChild(), new ArrayList<>(), 0, 0, ppr, depth + 1,
+                    deepening + 1, featureInteractions, memo, maxInteractionDepth, maxTreeDepth, maxDeepening, treeIndex);
+        }
+
+        String path = FeatureInteraction.interactionPathToStr(interactionPath, true, true);
+
+        FeatureInteraction foundFI = featureInteractions.get(featureInteraction.name);
+        if (foundFI == null) {
+            featureInteractions.put(featureInteraction.name, featureInteraction);
+            memo.add(path);
+        } else {
+            if (memo.contains(path)) {
+                return;
+            }
+            memo.add(path);
+            foundFI.gain += currentGain;
+            foundFI.cover += currentCover;
+            foundFI.fScore += 1;
+            foundFI.fScoreWeighted += pathProba;
+            foundFI.averageFScoreWeighted = foundFI.fScoreWeighted / foundFI.fScore;
+            foundFI.averageGain = foundFI.gain / foundFI.fScore;
+            foundFI.expectedGain += currentGain * pathProba;
+            foundFI.treeDepth += depth;
+            foundFI.averageTreeDepth = foundFI.treeDepth / foundFI.fScore;
+            foundFI.treeIndex += treeIndex;
+            foundFI.averageTreeIndex = foundFI.treeIndex / foundFI.fScore;
+            foundFI.splitValueHistogram.merge(featureInteraction.splitValueHistogram);
+        }
+
+        if (interactionPath.size() - 1 == maxInteractionDepth)
+            return;
+
+        foundFI = featureInteractions.get(featureInteraction.name);
+        SharedTreeNode leftChild = node.getLeftChild();
+        if (leftChild.isLeaf() && deepening == 0) {
+            foundFI.sumLeafValuesLeft += leftChild.getLeafValue();
+            foundFI.sumLeafCoversLeft += leftChild.getWeight();
+            foundFI.hasLeafStatistics = true;
+        }
+
+        SharedTreeNode rightChild = node.getRightChild();
+        if (rightChild.isLeaf() && deepening == 0) {
+            foundFI.sumLeafValuesRight += rightChild.getLeafValue();
+            foundFI.sumLeafCoversRight += rightChild.getWeight();
+            foundFI.hasLeafStatistics = true;
+        }
+
+        collectFeatureInteractions(leftChild, new ArrayList<>(interactionPath), currentGain, currentGain, ppl,
+                depth + 1, deepening, featureInteractions, memo, maxInteractionDepth, maxTreeDepth, maxDeepening, treeIndex);
+        collectFeatureInteractions(node.getRightChild(), new ArrayList<>(interactionPath), currentGain, currentGain, ppr,
+                depth + 1, deepening, featureInteractions, memo, maxInteractionDepth, maxTreeDepth, maxDeepening, treeIndex);
+    }
+    
+    public static TwoDimTable[][] getFeatureInteractionsTable(FeatureInteractions featureInteractions) {
+        TwoDimTable[][] table = new TwoDimTable[3][];
+        table[0] =  featureInteractions.getAsTable();
+        table[1] = new TwoDimTable[]{featureInteractions.getLeafStatisticsTable()};
+        table[2] = featureInteractions.getSplitValueHistograms();
+
+        return table;
     }
 }
